refactor : separate UpdateNoticePinBean from UpdateNoticeBean
